Refactor Commontator comment validation logic and move to Extensions module
